Add release ci for xcframework
