--- .github/workflows/apple-devices.yml
@@ -1,86 +1,0 @@
-# based on https://gist.github.com/learnopengles/60c945b92fbe60a0fecaa0143c35c4a for testing in simulator
-name: Apple Simulator devices
-
-on: pull_request
-
-env:
-  TOOLCHAIN: nightly
-  CARGO_TERM_COLOR: always
-  CARGO_DINGHY_REV: "121354f"
-
-jobs:
-  integration_tests:
-    # rust cross-compilation
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        rust-target: ["aarch64-apple-ios-sim" ]
-        apple-sim: [ "iOS" ]
-        os: [ macos-14 ]
-        include:
-          - rust-target: "aarch64-apple-tvos-sim"
-            os: macos-14
-            apple-sim: "tvOS"
-          - rust-target: "aarch64-apple-watchos-sim"
-            os: macos-14
-            apple-sim: "watchOS"
-    steps:
-      - uses: actions/checkout@v4
-      - uses: dtolnay/rust-toolchain@master
-        with:
-          toolchain: ${{ env.TOOLCHAIN }}
-          components: rust-src
-      - name: Cache
-        uses: actions/cache@v4
-        with:
-          path: |
-            ~/.cargo/bin
-            ~/.cargo/git
-            ~/.cargo/registry
-            target
-          key: ${{ github.workflow }}-${{ runner.os }}-${{ runner.arch }}-${{ github.job }}-toolchain-${{ env.TOOLCHAIN }}-cargo-${{ env.CARGO_DINGHY_REV }}-${{ hashFiles('**/Cargo.lock') }}
-      - name: Install cargo-dinghy
-        run: |
-          if ! command -v cargo-dinghy &> /dev/null; then
-            cargo install --git https://github.com/sonos/dinghy.git --force --rev ${CARGO_DINGHY_REV} cargo-dinghy
-          fi
-
-      - name: Start Simulator
-        uses: futureware-tech/simulator-action@v3
-        with:
-          os: ${{ matrix.apple-sim }}
-
-      - name: Build tests
-        env:
-          DINGHY_LOG: debug
-        run: |
-          cargo -Zbuild-std test --target ${{ matrix.rust-target }} --no-run
-
-      - name: Run Rust tests on simulator for ${{ matrix.rust-target }}
-        env:
-          DINGHY_LOG: debug
-        timeout-minutes: 20
-        run: |
-          cargo dinghy all-platforms
-          cargo dinghy all-devices
-          cargo -Zbuild-std test --target ${{ matrix.rust-target }}
-
-      - name: Set up Elixir
-        run: brew install elixir
-
-      - name: Install test server dependencies, compile and run in background
-        working-directory: ./tests/support/test_server
-        run: |
-          mix deps.get
-          mix compile
-          mix phx.server &
-
-      - name: Run Rust tests on simulator for ${{ matrix.rust-target }}
-        env:
-          DINGHY_LOG: debug
-        timeout-minutes: 20
-        run: |
-          cargo dinghy all-platforms
-          cargo dinghy all-devices
-          cargo -Zbuild-std test --target ${{ matrix.rust-target }} --features liveview-channels
-          cargo -Zbuild-std test --target ${{ matrix.rust-target }} --features liveview-channels-tls
