--- .github/workflows/macos.yml
@@ -1,67 +1,0 @@
-name: macos
-
-on: pull_request
-
-env:
-  CARGO_TERM_COLOR: always
-  TOOLCHAIN: nightly
-
-jobs:
-  integration_tests:
-    runs-on: macos-14
-
-    steps:
-      - uses: actions/checkout@v4
-      - name: Install Rust Nightly
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: ${{ env.TOOLCHAIN }}
-          default: true
-      - name: Cache
-        uses: actions/cache@v4
-        with:
-          path: |
-            ~/.cargo/bin
-            ~/.cargo/git
-            ~/.cargo/registry
-            target
-          key: ${{ github.workflow }}-${{ github.job }}-toolchain-${{ env.TOOLCHAIN }}-cargo-${{ hashFiles('**/Cargo.lock') }}
-
-# TODO: Fix this when enabling live-channels feature flag.
-#      - name: Cache deps
-#        id: cache-deps
-#        uses: actions/cache@v4
-#        env:
-#          cache-name: cache-elixir-deps
-#        with:
-#          path: tests/support/test_server/deps
-#          key: ${{ runner.os }}-mix-${{ env.cache-name }}-2-${{ hashFiles('**/mix.lock') }}
-#          restore-keys: |
-#            ${{ runner.os }}-mix-${{ env.cache-name }}-
-#      - name: Cache compiled build
-#        id: cache-build
-#        uses: actions/cache@v4
-#        env:
-#          cache-name: cache-compiled-build
-#        with:
-#          path: tests/support/test_server/_build
-#          key: ${{ runner.os }}-mix-${{ env.cache-name }}-2-${{ hashFiles('**/mix.lock') }}
-#          restore-keys: |
-#            ${{ runner.os }}-mix-${{ env.cache-name }}-
-#            ${{ runner.os }}-mix-
-#      - name: Set up Elixir, kotlin and ktlint
-#        run: brew install elixir kotlin ktlint
-#      - name: Install test server dependencies, compile and run in background
-#        working-directory: ./tests/support/test_server
-#        run: |
-#          mix deps.get
-#          mix compile
-#          mix phx.server &
-
-      - name: Get the JNA jar for testing kotlin.
-        run: wget 'https://repo1.maven.org/maven2/net/java/dev/jna/jna/5.14.0/jna-5.14.0.jar'
-      - name: Test
-        run: cargo test
-        env:
-          # This is required for the kotlin uniffi tests
-          CLASSPATH: "/Users/runner/work/liveview-native-core/liveview-native-core/jna-5.14.0.jar:/opt/homebrew/opt/kotlin/libexec/lib/kotlinx-coroutines-core-jvm.jar"
