+++ crates/core/liveview-native-core-swift/Sources/LiveViewNativeCore/.gitignore
@@ -1,2 +1,1 @@
 LiveViewNativeCore.swift
-PhoenixChannelsClient.swift
